Add documentation and other small fixes (#6)
* Fix issue when setting cuda architectures to all or native using env variables

* Add doc for building (in construction)

* Fix some issues in CMake cuda architecture setting

* Fix issue with scatter tail fitting factor

* Add help for building yrt-pet

* Update building.md

+++ yrt-pet/src/scatter/ScatterEstimator.cpp
@@ -126,8 +126,8 @@ namespace Scatter
 	float ScatterEstimator::computeTailFittingFactor()
 	{
 		std::cout << "Computing Tail-fit factor..." << std::endl;
-		float scatterSum = 0.0f;
-		float promptsSum = 0.0f;
+		double scatterSum = 0.0f;
+		double promptsSum = 0.0f;
 
 		for (bin_t bin = 0; bin < mp_scatterHisto->count(); bin++)
 		{
